requirements/common.txt
psycopg2~=2.7.1
python-magic~=0.4.13
elasticsearch~=2.4.1
pyquery~=1.2.17
raven~=6.1.0

django~=1.11.0
django-extensions~=1.9.1
django-taggit~=0.22.1
django-haystack~=2.6.0
django_compressor~=2.2.0
django_libsass~=0.7.0
django-sendfile~=0.3.11
django-allauth~=0.33.0
django-crispy-forms~=1.6.1
django-autoslug~=1.9.3
django-bootstrap-breadcrumbs~=0.8.2
django-appconf~=1.0.2

# API
djangorestframework~=3.6.2
markdown~=2.6.8
